tdf#167740 vcl: Fix Win IME inserting stale composition after dismissal
Related to tdf#155158: Previously, the Microsoft Pinyin IME failed to
send WM_IME_ENDCOMPOSITION when dismissed during composition. As a
result of this, LO became stuck in an IME input state (resulting in, for
example, an invisible cursor). The fix for this bug was to treat
dismissing the IME in this state as also ending composition.

Unfortunately, in the case of the Windows 10 Microsoft Korean IME, this
bug fix caused the last character to be repeated due to stale
composition state. This change updates the previous fix to also clear
this data prior to ending composition.

Change-Id: I5b2906c54af495e238a8335cc9c21cb42f6f3beb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188754
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>