tdf#130857 qt weld: Implement TextView scroll API
Implement weld::TextView API related to scrolling
in the Qt implementation, QtInstanceTextView.

For the license dialog newly supported since

    Change-Id: I1441119bd334a36dd260d491347575093cd98aa0
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Aug 1 12:34:22 2025 +0200

        tdf#130857 qt weld: Support extension license dialog

, this causes the "Scroll Down" button to be greyed
out as expected when using the scrollbar to move
down to the end of the license document.

Clicking the "Scroll Down" button moves the license
text down by one page as expected according to the
implementation in LicenseDialogImpl::PageDown and
what is seen with the gtk3 and gen VCL plugins.

Change-Id: I803d1f0f9108fc20836eab3ffea164e39be2d3fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188769
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>