tdf#130857 svtools: Better detect non-user values in File Services dlg
PlaceEditDialog::m_bLabelChanged is supposed to
indicate whether the value in the "Label" entry
in the file services dialog was entered by the user.

Therefore, reset to false for both branches in
PlaceEditDialog::EditHdl where a text is set
programmatically, not just for one.

While at it, also connect the signal handlers in
the ctor after manually setting various values.

While the GTK and VCL implementation suppress
signals when values are set programmatically
(s. e.g. SalInstanceEntry::set_text), the Qt
implementation doesn't do that (yet?), so
the label would no longer update according to
the currently selected service for the following
scenario without this commit in place for the
qt6 VCL plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1
with an upcoming commit that declares support for
using native Qt widgets for that dialog:

* "File" -> "Open Remote"
* "Manage Services" -> "Add service"
* in the "Type" combobox, select "WebDAV"
* in the "Type" combobox, select "SSH"

Without this commit, the "Label" would still
be set to "WebDAV", now it is "SSH" as expected.

Change-Id: I5d0a1b8a5ad3ff1d5bd073e7e57692bfdd3ea35b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188776
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>