related tdf#167770 docx import/export: support INSIDE/OUTSIDE
During code reading, I noticed a do-nothing nWidth.

It became obvious that this actually did nothing with
commit 5b5995c05a17b1b16617ef8b477a657ce2a9406c
Author: Noel Grandin on Thu Jul 20 10:15:07 2017 +0200
    loplugin:unusedfields in writerfilter

Everything about that emulation looked completely wrong,
so I've changed it so that it maps
to LibreOffice's INSIDE/OUTSIDE - as vmlshape already does.

Inside and Outside are essentially identical,
so they should be emulated identically as well.

Inside/Outside react the same way to "margin" as to "page",
(and those are the only two choices)
so both of those should also be emulated identically.

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167770_marginInsideOutside

Change-Id: If2ef6af13f4a4acf85d8ed981880fe631557ce9d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188834
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>