tdf#130857 sw: Define widgets in footnote/endnote dlg in order
Switch the order in which the widgets in the .ui file
are defined so that the order matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

This dialog can be triggered using "Tools" -> "Footnote/Endnote Settings"
in Writer.

Change-Id: I376bc4a47cbf3eaaa5f9f9a3e411fd1af9d21502
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188830
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins