sw a11y: Set mnemonic widgets in footnote/endnote dialog
Set the mnemonic widgets for the labels, which makes
moving focus to them pressing the corresponding
accelarator/mnemonic key work and also ensures that
the label-for/labelled-by a11y relation is set, making
e.g. the Orca screen reader on Linux announce the
corresponding label when a control receives focus,
so it becomes clear what can be configured there.

This dialog can be triggered using "Tools" -> "Footnote/Endnote Settings"
in Writer.

Change-Id: I4d67132a430d5547f294887509ddad2c8592a1c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188832
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>