Fix HarfBuzz build breakage when building on macOS Tahoe
Starting in Xcode 26, compiling with -D_LIBCPP_ENABLE_ASSERTIONS will
fail due to a change made by Apple.

HarfBuzz uses meson to define the macros needed to build HarfBuzz
and meson 1.7.2 and 1.8.0 still emit -D_LIBCPP_ENABLE_ASSERTIONS.
So upgrade the internal meson to 1.8.3 (i.e. the first version that
is compatible with Xcode 26) and use the internal meson if building
with Xcode 26 or higher and the external meson version is too old.

Change-Id: I0d3545b0b1a22156179fec1382260adca1e9221c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188863
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins