tdf#130857 qt weld: Support "Enter Master Password" dialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

The dialog can be triggered as follows:

* "Tools" -> "Options"
* "LibreOfficDev" -> "Security"
* enable the "Persistenly save passwords for web connections"
  checkbox
* type the same password twice and confirm to close the dialog
* click the "Master Password" button

When entering the wrong password, another instance
of the dialog is currently shown right away, hiding
the message dialog telling that the initially entered
password is wrong. This will be addressed in a separate
commit.

Change-Id: Ibf12c9450902058271559b65b1307d4792901485
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188877
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>