SetOutputSizePixel(bAlphaMaskTransparent=true) should make..
the background transparent.

I initially tried debugging this on both Windwos and macOS.
Both quickly descended into nightmare.
On macOS we have CGLayer interposing itself with a bunch
of extra weirdness.
On Windows we have WinSalVirtualDevice sometimes rendering
to a GDI bitmap and then inexplicably switching to using
a skia surface.

I will leave all that alone for now and just use a bigger
hammer.

Change-Id: I0a485b526af29e841e7267430800bd8ca394d871
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188869
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins