tdf#130857 qt weld: Support Java "Class Path" dialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

The dialog can be triggered as follows:

* "Tools" -> "Options"
* "LibreOfficDev" -> "Advanced"
* select a JVM in the treeview
* click the "Class Path" button

Currently, both the icon and the path are displayed
in the first column of the tree view and the second
column remains empty. This will be addressed in
upcoming commits.

Change-Id: Id94e22e879f712bd967431ec0f9dc34e4480e3dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188890
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins