tdf#130857 qt weld: Evaluate GtkCellRenderers to know supported types
So far, The GtkCellRenderer objects in .ui files were ignored
by QtBuilder.
However, for the automatic detection of what is the first column
of a specific type (e.g. the first text column), that information
is relevant, and e.g. the GTK implementation in GtkInstanceTreeView
sets `GtkInstanceTreeView::m_nTextCol` based on that, then uses
it when inserting text via the weld::TreeView::insert method.

Extend the logic in QtBuilder and QtInstanceTreeView to also
evaluate the cell renderers set in the .ui file and store the
item data roles supported for each column based on those renderers.

The Qt data model allows setting any of the roles exactly once for
each item, so there is generally no need to explicitly enable
something. But in order to adhere to the weld::TreeView logic/API,
that kind of information is relevant.

Let QtBuilder store a list of supported roles for each column
(i.e. use a list of lists for all columns) in a property set
for the QTreeView, and then use that property in the
QtInstanceTreeView ctor to set the newly introduced
QtInstanceTreeView::m_pColumnRoles member.

Adjust QtInstanceTreeView::firstTextColumnModelIndex to
make use of those stored roles, to reliably detect what
is the first text column according to the GTK/weld::TreeView
model, without requiring any entries to have been inserted
previously.

This is one step towards addressing the issue mentioned in
previous commit

    Change-Id: Id94e22e879f712bd967431ec0f9dc34e4480e3dd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sun Aug 3 22:15:20 2025 +0200

        tdf#130857 qt weld: Support Java "Class Path" dialog

> Currently, both the icon and the path are displayed
> in the first column of the tree view and the second
> column remains empty. This will be addressed in
> upcoming commits.

Change-Id: Ic8a59b254f3bac14c3b812f00d0cbfa3740432f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188891
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins