tdf#130857 qt weld: Unbreak "Writing Aids" option page
Prevent trying to use native Qt widgets for the
"Languages and Locales" -> "Writing Aids" page as
long as the dialog itself isn't supported yet, as that
results in that page looking broken with when using the
qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Declare that .ui file only as supported when a native
parent is used.

That file was previously declared as supported in

    commit 66a674f915badf9e586b8e3854954d71dc908a4d
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Feb 21 15:33:53 2025 +0100

        tdf#130857 qt weld: Support spelling options dialog

because in that scenario it does have native parent
and works fine there.

Change-Id: I185827e0931245d87424dea71fecfd86ca627236
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188901
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>