tdf#130857 qt weld: Unbreak "Print" option page
Prevent trying to use native Qt widgets for the
"LibreOfficeDev Writer" -> "Print" page in the
"Tools" -> "Options" dialog long as the dialog itself
isn't supported yet, as that results in that page
looking broken with when using the qt5 or qt6 VCL
plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Declare that .ui file only as supported when a native
parent is used.

Change-Id: Ifbd7e30f162fd3bed6426a145ef36827a1df401a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188902
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>