tdf#130857 qt weld: Ignore GtkCellRenderer without tree view parent
No more assert that all GtkCellRenderers have
a GtkTreeView ancestor (i.e. are children of a
GtkTreeViewColumn that is a child of a GtkTreeView).

That assert was introduced in

    commit 8e33908b538c7700be6f766c472f069d7f4dd15b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sun Aug 3 22:17:34 2025 +0200

        tdf#130857 qt weld: Evaluate GtkCellRenderers to know supported types

However, GtkCellRenderers can also be used as children
for GtkComboboxText as is the case e.g. in the Spelling
Dialog (UI file: cui/uiconfig/ui/spellingdialog.ui) for
the text language combobox that displays an icon in
addition to the text.

QtInstanceComboBox doesn't need the model/cell renderer
information for that one, so simply return early again
instead of triggering an assert, as used to be the case before
the above-mentioned commit.

Also return from QtBuilder::makeObject after
running QtBuilder::enableTreeViewColumnDataRole
as there is nothing else to do in case of the cell
renderers.

(Assert seen getting triggered in a WIP branch for declaring
support for the spelling dialog.)

Change-Id: Id35a4d1f6c9a9728b547a5c0103502a89e4a8b9d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188911
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>