Allow SAL_LOG_FILE to add process ID to file name
...so that different processes (e.g., when soffice.bin gets automatically
restarted) less easily overwrite each other's log files.  The protocol (to
mostly stay backwards compatible and only use a single file name by default) is
to append a process ID iff the value of SAL_LOG_FILE ends in a hyphen, in which
case the process ID is appended to that value.

Change-Id: I516d7a286d029a16afc2f1379d5353f3f0eb5767
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188914
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <stephan.bergmann@collabora.com>