Make FontId of WinFontFace monothonically increasing
It was generated from the position in font collection; that meant,
that after rebuilding the collection, the same id could refer to
a different font. (The collection gets refreshed when a document
is loaded with embedded fonts; or when a user installs fonts.)

The ids are used as keys in a blob cache (WinFontFace::GetHbTable);
and that cache isn't rebuilt when the collection is refreshed; it
could make a wrong cache entry be used.

This fix is from code reading; I don't have a repro scenario. It
is unlikely that this would overflow even on 32-bit architecture;
and then, it would be no worse than previously.

Change-Id: Ifd09cdbcb98b7f739da871fbe47ce1d107f0e886
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188916
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins