tdf#163802 a11y: Fully initialize SpellDialog before showing
SpellDialog can't do all of its initialization in the
ctor because it needs to call virtual methods on the
SpellDialogChildWindow* that gets passed as a param,
but that one calls the SpellDialog from its ctor, i.e.
it hasn't been fully initialized yet.

For that reason, the part of the initialization done in
InitHdl was delayed by using a user event (posted via
Application::PostUserEvent).
This also includes setting the dialog title based on
the selected text language.

As a consequence of this, the title, and therefore the
accessible name of the dialog would change after it was
shown, which would result in Orca speech output for the
dialog getting interrupted (see tdf#163802).

Avoid the problem by not depending on the next user
event getting processed, but calling the method
that does the rest of the initialization manually
after the SpellDialogChildWindow constructor has
finished.

In order to do that, add a virtual SfxChildWindow::Initialize
(that does nothing) and override it in the
SpellDialogChildWindow subclass.

Since SpellDialog::Initialize (renamed from SpellDialog::InitHdl)
is now guaranteed to be called before the dialog shows
and that one always sets a title (by calling
UpdateBoxes_Impl), drop the no longer needed

    // fdo#68794 set initial title for cases where no text has been processed
    // yet to show its language attributes
    OUString sTitle = rParent.HasGrammarChecking() ? m_sTitleSpellingGrammar : m_sTitleSpelling;
    m_xDialog->set_title(m_xDialog->strip_mnemonic(sTitle.replaceFirst("$LANGUAGE ($LOCATION)", "")));

from the ctor.

Additional fix for GTK 3 to avoid incorrect
object:property-change:accessible-name events from
getting sent when the title didn't actually change [1]:

    commit 0cc9e603f4635c31d2a8a5f602869474d329e728
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Aug 4 15:31:06 2025 +0200

        a11y: Only notify when window title changed

        In gtk_window_set_title_internal, return early
        when called with the same title as is already set.

        This also ensures that notifying of a change
        of PROP_TITLE only happens when the property
        actually changed, which also prevents invalid
        object:property-change:accessible-name AT-SPI
        events from getting emitted on the accessibility
        layer when the title hasn't actually changed.

        Fixes: #7690
        Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8814>

Output of the pyatspi listener script from tdf#163802,
attachment 197494 for gtk3 before this LO commit (note the
object:property-change:accessible-name event after
the object:state-changed:showing event):

    event type: object:property-change:accessible-name, event source: Spelling: , states: ['ENABLED', 'SENSITIVE', 'SHOWING', 'VISIBLE']
    object:property-change:accessible-name(0, 0, Spelling: )
            source: [dialog | Spelling: ]
            host_application: [application | soffice]
            sender: [application | soffice]

    event type: object:state-changed:showing, event source: Spelling: , states: ['ENABLED', 'SENSITIVE', 'SHOWING', 'VISIBLE']
    object:state-changed:showing(1, 0, 0)
            source: [dialog | Spelling: ]
            host_application: [application | soffice]
            sender: [application | soffice]

    event type: object:property-change:accessible-name, event source: Spelling: English (UK), states: ['ENABLED', 'SENSITIVE', 'SHOWING', 'VISIBLE']
    object:property-change:accessible-name(0, 0, Spelling: English (UK))
            source: [dialog | Spelling: English (UK)]
            host_application: [application | soffice]
            sender: [application | soffice]

Output with this change in place:

    event type: object:property-change:accessible-name, event source: Spelling: English (UK), states: ['ENABLED', 'SENSITIVE', 'SHOWING', 'VISIBLE']
    object:property-change:accessible-name(0, 0, Spelling: English (UK))
            source: [dialog | Spelling: English (UK)]
            host_application: [application | soffice]
            sender: [application | soffice]

    event type: object:state-changed:showing, event source: Spelling: English (UK), states: ['ENABLED', 'SENSITIVE', 'SHOWING', 'VISIBLE']
    object:state-changed:showing(1, 0, 0)
            source: [dialog | Spelling: English (UK)]
            host_application: [application | soffice]
            sender: [application | soffice]

[1] https://gitlab.gnome.org/GNOME/gtk/-/commit/0cc9e603f4635c31d2a8a5f602869474d329e728

Change-Id: I6cdfe8443ae435e1e644b78942cf63e9df1d400f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188937
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins