tdf#130857 Define widgets in spelling dialog in order
Switch the order in which the widgets in the .ui file
are defined so that the order matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

The dialog can be triggered as follows:

* start Writer
* type some random text
* "Tools" -> "Spelling"

The dialog is not yet declared as supported for use with
native Qt widgets, but in a WIP branch it makes moving
focus with the Tab key work as expected.

Change-Id: I3a8d750a543deeab45ee0085088f4f90f32307a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188938
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins