ScriptForge new session.GetRangeFromCalc method
The
   GetRangeFromCalc(filename: str, range: str)
method (stored in the Session service)
reads the values contained in the given
range located in the given Calc file.

The file does not need to be open. This
makes this shortcut particularly useful.

The content of the range is returned either
- as a scalar for a single cell
- as a 1D-array or a tuple when the range
  is contained in a single row or column
- as a 2D-array or a tuple of tuples for
  a rectangular range of adjacent cells.

Clean error management is included in the code.

The method is available for Basic and Python
user scripts.

The user documentation needs to be updated.

Change-Id: I39cd218f0bf3a6a21e3ada42826069eeee451c6e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189000
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins