tdf#167658 vcl: Honor GtkIconView::columns + calc size
For SalInstanceIconView, i.e. the vcl implementation
of weld::IconView, using a vcl IconView control,
take the GtkIconView::columns [1] property into
account:

Introduce IconView::m_nFixedColumnCount as a new
member for which VclBuilder sets the value accordingly
when processing a .ui file.

Override the SvTreeListBox::GetOptimalSize base class
implementation for IconView to instead return the
size required when all icons are arranged using the
given amount of columns if one was explicitly set.
Otherwise, keep using the logic used so far.

This e.g. prevents the IconViews in Writer's "Format"
-> "Page Style" dialog, tab "Borders" from occupying a
lot of unused vertical space (as the "columns" property
is already set for those in the UI file,
cui/uiconfig/ui/borderpage.ui) and thus addresses
that aspect mentioned in previous commit

    Change-Id: I6d8d8d5ac38d7ebb62c0d9963ec2125d417a3a49
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Aug 7 10:17:57 2025 +0200

        tdf#167658 Update vcl IconView entry size when inserting item

        When inserting a new entry into a SalInstanceIconView,
        using SalInstanceIconView::insert, let the IconView
        control recalculate an appropriate entry size, as
        is already done when the image of an entry is updated
        in SalInstanceIconView::set_image since

            commit 9f2fc6fbf74cbc0643f380aa4b799e36997734aa
            Date:   Sat Mar 8 17:29:01 2025 +0530

                tdf#165610 sd: master slide panel is broken with "show large preview"

        However, (in both cases) only do so if no custom entry
        width was explicitly set using SalInstanceIconView::set_item_width.

        This makes the presets in Writer's "Format" -> "Page Style" dialog,
        tab "Borders" show again with the expected size and in one row
        instead of one item per row.
        The IconView still has a lot of unused space, which will be addressed
        in a separate commit.

        Change-Id: I6d8d8d5ac38d7ebb62c0d9963ec2125d417a3a49

[1] https://docs.gtk.org/gtk3/property.IconView.columns.html

Change-Id: I95e13d748c4c178b7492665bbc9e1a5980c7ef81
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189086
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>