tdf#167658 Update vcl IconView entry size when inserting item
When inserting a new entry into a SalInstanceIconView,
using SalInstanceIconView::insert, let the IconView
control recalculate an appropriate entry size, as
is already done when the image of an entry is updated
in SalInstanceIconView::set_image since

    commit 9f2fc6fbf74cbc0643f380aa4b799e36997734aa
    Date:   Sat Mar 8 17:29:01 2025 +0530

        tdf#165610 sd: master slide panel is broken with "show large preview"

However, (in both cases) only do so if no custom entry
width was explicitly set using SalInstanceIconView::set_item_width.

This makes the presets in Writer's "Format" -> "Page Style" dialog,
tab "Borders" show again with the expected size and in one row
instead of one item per row.
The IconView still has a lot of unused space, which will be addressed
in a separate commit.

Change-Id: I6d8d8d5ac38d7ebb62c0d9963ec2125d417a3a49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189080