cid#1642989 Wait without held lock
restore the lock at the end of SerfLockStore::stopTicker

SerfLockStore::removeLockImpl calls this, and that is called
by SerfLockStore::refreshLocks in a loop, so if it was unlocked
at the end of the first loop, then its still unlocked on a 2nd
loop.

Probably doesn't matter as there's only a 2nd unlock attempt
if m_aLockInfoMap.empty() which presumably can only happen once.

Change-Id: Iede91b1b4ef9368bcf4e62cb85bf5a0c745e0ba7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189058
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins