tdf#130857 svx: Don't use GtkBox as GtkFrame label child
The weld::Frame API implementations expect that
the `<child type="label">` child of a GtkFrame
object in a UI file is a GtkLabel.

In addition, QtBuilder also expects that when
processing the UI file as it doesn't use a
separate widget, but sets the label's text directly
for the QGroupBox (that's used for the GtkFrame
object in the .ui file).

The "Header"/"Footer" tab pages used e.g. in Writers
"Format" -> "Page Style" dialog however were using
a GtkBox child, containing 2 labels: One having the
string to be used for the "Header" dialog, one the
string to be used for the "Footer" dialog set as text,
and only that label was set to visible in the SvxHFPage
ctor.

Replace that approach by using a GtkLabel child,
and calling weld::Frame::set_label with the corresponding
string instead.

Change-Id: I86754be7c0ce068ceb6216fd0463657b05f4ca41
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189111
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins