tdf#130857 sw mailmerge: Use GtkBox instead of GtkFrame
I cannot see any reason why this was using a GtkFrame
and the GtkBox containing the "Save as single document"
and "Save as individual documents" radiobuttons was set
as the GtkFrame's label child. This breaks the assumption
that the label child is always a GtkLabel object, see

    commit 82308d6a6853463f660dc6d6a11a2c12d198c4ed
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Aug 7 16:15:12 2025 +0200

        tdf#130857 svx: Don't use GtkBox as GtkFrame label child

for more background.

Replace that unusual approach by using a GtkBox
instead of the GtkFrame, and simply have the
GtkBox containing the radio buttons as the
first child of the GtkBox.

No change in behavior seen in a quick test.

The dialog can be triggered as follows:

* start Writer
* "Insert" -> "Field" -> "More Fields"
* switch to "Database" tab
* select "Mail merge fields" for the type
* click "Browse" button next to the "Add database file"
  label and select any database file
* in the "Database Selection" treeview, select any column
  of the database
* press the "Insert" button
* close the "Fields" dialog
* press Ctrl+P to trigger the print dialog
* in the dialog asking whether to print a form letter,
  press the "Yes" button

Change-Id: I49b57c556ab81045e199ac0579bc886f0654c520
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189148
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>