tdf#130857 sw: Don't use GtkBox as GtkFrame label child
The "typeft" label was never shown, but only used to
get the "Type" string in order to set it as the text
for the "levelft" label in certain circumstances.

Change the approach and use the already existing
string resource STR_AUTH_FIELD_AUTHORITY_TYPE.
That allows to drop the unused label, which in
turn allows dropping the GtkBox previously containing
both labels, which was set as the "label" child
of the GtkFrame. Instead, set the "levelft" label
directly as the label child, which is more in
line with weld::Frame API expectations.

The dialog can be triggered from Writer using
"Insert" -> "Table of Contents and Index..."
-> "Table of Contents, Index or Bibliography..."

This commit is similar to previous commit

    commit 82308d6a6853463f660dc6d6a11a2c12d198c4ed
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Aug 7 16:15:12 2025 +0200

        tdf#130857 svx: Don't use GtkBox as GtkFrame label child

See the commit message of the above-mentioned commit for more
background.

Change-Id: I88340d4f00dbe3b872f5547612694f0854f17238
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189151
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>