vcl: Add Window::GetTypeName to expose type to name conversion
Many times it is useful when to log the window type name instead
of only the number of the type enum (i.e. when you check the list
of parent windows). This change adds a new method that exposes the
type to type name conversion that is already present in the code
for the Window.

Change-Id: Ice97f7ad27988d99f0e45051e0d8e44befe955e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189162
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins