tdf#130857 sw: Set frame's label directly, not via label child
In order to set the title/label for the frame, do
it directly using the weld::Frame, not a weld::Label
object for its label.

The native Qt implementation for weld::Frame uses a
QGroupBox and doesn't use a separate label child, so
this commit prepares for adding support for that
dialog for that one at some point.

The dialog can be triggered from Writer using
"Insert" -> "Table of Contents and Index..."
-> "Table of Contents, Index or Bibliography...";
the "Entries" page is the relevant one.

Change-Id: I8c13ad44383a3bacf11db0f75a5c241c9fb4a299
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189170
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>