tdf#130857 formula: Don't have getter for Label, but text
Replace RefEdit::GetLabelWidgetForShrinkMode that
returns a weld::Label with RefEdit::GetLabelTextForShrinkMode
that only returns the label's text.

This is the only information of interest in the caller, and
for the case where the label is actually the label child of
a frame (as in the example mentioned below) instead of an
independent label, there will be no actual label for the
native Qt implementation, as the QGroupBox used in
QtInstanceFrame handles its title/text itself.

This is one step towards getting rid of
weld::Frame::weld_label_widget for the above reason.

One way to trigger this code path:

* start Calc
* select some cells
* "Data" -> "Define Range"
* type "foobar" into the name field
* click on the "Shrink" button

Change-Id: I3b987a052fe4ba72a6793671a7a81142b12afe95
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189217
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>