tdf#167721 writerfilter: ensure w:left applied if w:leftChars=0
Third problem - if leftChars is disabled,
then w:left needs to be applied to the paragraph.

If BOTH were defined as direct properties
  <ind w:left="1234" w:leftChars="100"/>
then xTextAppend->finishParagraph has applied both
of these properties to the paragraph,
and it is somewhat ambiguous which one exists.
(In my testing, PROP_PARA_LEFT_MARGIN_UNIT wins).

So in this section of code, we are reviewing the situation
and re-applying the appropriate property.
In the special case when the disabled-leftChars property
was written to the actual paragraph,
the direct-or-inherited w:left needs to be
rewritten onto the paragraph.

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits2

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits3

Change-Id: Ic36db6d4934876d800770fe8f3303c7eca06d532
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189234
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189549