tdf#130857 formula: No longer remember the reference label widget in RefEdit
No longer store the weld::Label* passed to
RefEdit::SetReferences as a class member RefEdit::mpLabelWidget,
but instead have a std::function that returns the
label's label/text.

No change in behavior intended yet, but this prepares
for allowing non-weld::Label widgets to be used in
the future.

See also previous commit

    Change-Id: I3b987a052fe4ba72a6793671a7a81142b12afe95
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Aug 8 10:54:27 2025 +0200

        tdf#130857 formula: Don't have getter for Label, but text

for more background.

Change-Id: I45f6c328b7da134ca0d4fcb4b078a50056c6564a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189219
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins