tdf#167721 writerfilter styles: use provided w:left when leftChars=0
Fourth Problem - styles need to apply either the value of w:leftChars
or the value of w:left, but not both.

If w:leftChars=0, then the style gets the value of w:left
(either an implied 0, or else specified, but NOT inherited!)

Otherwise, leftChars (which might be inherited) takes priority.

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits2

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits3

Change-Id: Ia19e48f57c6a716e3275df6261e7fb257f0cd788
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188930
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>