Adjust bounds of native controls for macOS Tahoe
macOS Tahoe increased the width of push buttons, listboxes, and
spin buttons. macOS Taho *only* uses the new sizes if the code is
built with Xcode 26 or higher. The same code built on earlier
versions of Xcode will display native controls with the pre-Tahoe
sizes. Hence, the use of the following unwieldy style:

 #if __MAC_OS_X_VERSION_MAX_ALLOWED >= 260000
     if (@available(macOS 26, *))
         ; // Do macOS Tahoe or later code
     else
 #endif
         ; // Do pre-macOS Tahoe or earlier code

Eventually, when official builds are being done on macOS Tahoe,
but still supporting pre-macOS Tahoe versions, the #if and #endif
can be removed:

     if (@available(macOS 26, *))
         ; // Do macOS Tahoe or later code
     else
         ; // Do pre-macOS Tahoe or earlier code

Also, adjusted the width of flexible push buttons and spin buttons
so that the focus ring is not clipped on pre-macOS Tahoe versions.

Change-Id: I2f86f73590a50751dc25000dfc1e9cccb4f4a962
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189282
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins