tdf#160276: CUI: Increase left margin for comment author and date labels
The author and date labels in the comment sidebar were flush with the
left edge, making them feel cramped.

This change adds a 2px start margin to the 'author' and 'date' GtkLabels
in sw/uiconfig/swriter/ui/annotation.ui to give them some breathing
room and improve the visual layout.
The initial suggestion was to align these labels with the main comment
text. However, as this metadata is of secondary importance and doesn't
need to strictly align, the simpler solution of adding a small left
margin was chosen.

Change-Id: I21892575ee393e96daa675ccc95fb0bb05efed9b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189285
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>