tdf#162501: SD: Update list of drawing scale factors
The predefined list of drawing scales in Draw/Impress was missing
common factors used for imperial/English units, such as 1:12.

As per the design team's recommendation in the bug report, this
patch updates the hardcoded list to include these factors while
maintaining a concise set of useful metric scales.
The new list is: 1, 2, 5, 10, 12, 24, 48, 50, 100.

Additionally, the array size in sd/source/ui/app/scalectrl.cxx
was refactored from a magic number to use a TABLE_COUNT constant,
matching the pattern in tpoption.cxx for improved maintainability.

Change-Id: I1c2610fa218233f320207594ba1086a749c636be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189309
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>