Keep track of documents' embedded fonts, and unregister them on unload
First, not releasing the documents could degrade performance in some
long-running instances.

Also, this will help when we implement support for restricted embedded
fonts, which must not show in font lists, when the document was closed.

EmbeddedFontsHelper now keeps track of all the embedded fonts, using a
static map s_EmbeddedFonts; each font has a refcount, and unregistered
only when its refcount is zero. SfxMedium keeps track of its own fonts;
and in its dtor, it calls EmbeddedFontsHelper::releaseFonts.

Change-Id: I0afe971a785c4709cc54a3fab98953db96dbd05f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189347
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>