tdf#130857 qt weld: Implement QtInstanceDialog::{,undo_}collapse
Implement the weld::Dialog::collapse and
weld::Dialog::undo_collapse API that can be used
to temporarily hide all dialog widgets except for
one or two, and restore previous visibilibty of
all widgets in weld::Dialog::undo_collapse.

This is used e.g. in Calc's "Data" -> "Define Range"
dialog, for which support will be declared in an
upcoming commit.

Change-Id: I47362242a918c2a0d7b7c8658b376b4cf28567db
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189356
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins