tdf#130857 qt weld: Support sc "Define Range" dialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog makes use of the logic newly implemented
for QtInstanceDialog in previous commit

    Change-Id: I47362242a918c2a0d7b7c8658b376b4cf28567db
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Aug 11 11:40:53 2025 +0200

        tdf#130857 qt weld: Implement QtInstanceDialog::{,undo_}collapse

To trigger that code path:

* start Calc
* select some cells
* "Data" -> "Define Range"
* type "foobar" into the name field
* click on the "Shrink" button
* select some range in the document
* click the "Expand" button

Other than for the gtk3 or gen/VCL implementations, the "Range"
label of the corresponding "GtkFrame" object defined in the
UI file (sc/uiconfig/scalc/ui/definedatabaserangedialog.ui)
remains visible when the dialog gets collapsed, because the
text is set in the QGroupBox that is used for the "GtkFrame"
object in the Qt weld implementation and which is the parent
of the edit that remains visible together with all of its parents,
while for the GTK and VCL implementations, the label is a separate
widget, and therefore not a parent widget of any of the widgets
that should remain visible.

Change-Id: If1ceb6a744f267f69325c7ef8df5907dc98a9554
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189357
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>